Use embedded worker entrypoint in wasm during development
