+++ web/playground/wasm/src/util/mod.rs
@@ -1,5 +1,9 @@
+mod is_dev;
+mod or_throw;
 mod send_wrapper;
 mod variadic_function;
 
+pub use is_dev::*;
+pub use or_throw::*;
 pub use send_wrapper::*;
 pub use variadic_function::*;
